Implement file downloading on server

+++ src/main/java/ru/yandex/practicum/catsgram/service/PostService.java
@@ -77,7 +77,7 @@ public class PostService {
         throw new NotFoundException("Пост с id = " + newPost.getId() + " не найден");
     }
 
-    public Post getPostById(int postId) {
+    public Post getPostById(long postId) {
         return posts.values()
                 .stream()
                 .filter((post) -> post.getId() == postId)
